adding more functionality

+++ comite-olimpico/src/main/java/com/cenfotec/segundoexamenparcial/builder/AthleteDetailsBuilder.java
@@ -1,6 +1,5 @@
 package com.cenfotec.segundoexamenparcial.builder;
 
-import com.cenfotec.segundoexamenparcial.domain.AthleteDetails;
 import com.cenfotec.segundoexamenparcial.domain.BodyMassIndex;
 
 import java.util.List;
@@ -8,7 +7,7 @@ import java.util.List;
 public interface AthleteDetailsBuilder {
 
     AthleteDetailsBuilderImpl sport(String sport);
-    AthleteDetailsBuilderImpl height(Long height);
+    AthleteDetailsBuilderImpl height(Double height);
     AthleteDetailsBuilderImpl weight(Long weight);
     AthleteDetailsBuilderImpl gender(String gender);
     AthleteDetailsBuilderImpl  bodyMassIndexArrayList (List<BodyMassIndex> bodyMassIndexArrayList);
